requirements/base.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-ratelimit>=4.1
django-axes>=6.4
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
gunicorn>=21.0
django-extensions>=3.2
dj-database-url>=2.1
